Add files via upload
We undertook a challenge that revolved around the devastating impact of Hurricane Maria, which struck Puerto Rico in 2017. This challenge involved analyzing pre- and post-event satellite images of the entire map of San Juan, Puerto Rico. We used an object detection model that reached a MAP score around 0.75 and helped to identify 4 types of buildings: Damaged Residential, Damaged Commercial, Undamaged Residential, and Undamaged Commercial.
    After experimenting with two methods, we opted for using a larger dataset despite its lower precision, prioritizing scalability and a higher MAP score to address widespread impact. The model faced challenges with class imbalance, which negatively affected the detection of damaged buildings. Our findings are applicable in various domains such as infrastructure planning, durable construction, area zoning, insurance, disaster simulations, and historical analyses.